+++ CHANGELOG.md
@@ -6,7 +6,11 @@ All significant changes to this software be documented in this file.
 
 ### Breaking changes
 
-* Rename `BSize::with` to `BSize::map`. A new `BSize::with` method has been added that accepts a closure returning arbitrary type, allowing for mapping a `BSize` to any other type. This follows `LocalKey::with`'s naming conventions.
+* Renamed `BSize::with` to `BSize::map`. A new `BSize::with` method has been added that accepts a closure returning arbitrary type, allowing for mapping a `BSize` to any other type. This follows `LocalKey::with`'s naming conventions.
+
+### New features
+
+* Added `nightly` feature for using `BSize` with nightly-only features like `const_ops` and `const_trait_impl`.
 
 ## v0.2.1 (2026-06-27)